fix(navigation): address review comments on the deep link gate
Rethrow cancellation instead of swallowing it: the navigate blocks and the
splash check now use the project's resultCatching, which calls ensureActive()
before turning a failure into a Result, so a cancelled coroutine propagates
rather than being logged or reported as an unknown splash state.

Keep external input out of the logs. Deep links reach navigateFromUri from
outside the app and carry a trade, channel or profile id after the route, and
the log lines only need the route, so the id is dropped. The trade-not-found
warnings in both presenters log a short id, matching shortTradeId elsewhere.

Reset the chat-derived state with the rest on re-initialisation. The chat row
renders as soon as the trade is published, which happens before its collector
runs, so msgCount and lastChatMsg would carry the previous trade's numbers.
TradeChatPresenter also drops its delivery-status observers there: cancelling
the job stops the collector but leaves the observers on the previous trade's
messages, and onViewUnattaching only runs when the screen leaves.

Assert the superseded deep link is dropped rather than merely counting one
navigation: the test now stubs a distinct NavUri per uri, so a regression that
opened the first link would fail instead of passing.

Consolidate the four equivalent private FakeTradesServiceFacade copies into one
test util, parameterised on the only member that varied between them.
