+++ shared/presentation/src/androidUnitTest/kotlin/network/bisq/mobile/presentation/offer/take_offer/TakeOfferCoordinatorTest.kt
@@ -24,6 +24,7 @@ import network.bisq.mobile.domain.model.trade.TradeRoleFilter
 import network.bisq.mobile.domain.model.trade.TradeSort
 import network.bisq.mobile.presentation.common.test_utils.FakeConfigServiceFacade
 import network.bisq.mobile.presentation.common.test_utils.FakeMarketPriceServiceFacade
+import network.bisq.mobile.presentation.common.test_utils.FakeTradesServiceFacade
 import network.bisq.mobile.presentation.common.test_utils.OfferTestFactory
 import network.bisq.mobile.test.mocks.SettingsRepositoryMock
 import network.bisq.mobile.test.presentation.coroutines.PlatformPresentationKoinTestBase
@@ -34,57 +35,6 @@ import kotlin.test.assertTrue
 
 @OptIn(ExperimentalCoroutinesApi::class)
 class TakeOfferCoordinatorTest : PlatformPresentationKoinTestBase() {
-    private class FakeTradesServiceFacade(
-        private val takeOfferResult: Result<String> = Result.success("trade-1"),
-    ) : TradesServiceFacade {
-        override val selectedTrade: StateFlow<TradeItemPresentationModel?> = MutableStateFlow(null)
-        override val openTradeItems: StateFlow<List<TradeItemPresentationModel>> = MutableStateFlow(emptyList())
-        override val closedTradesChangeTick: StateFlow<Int> = MutableStateFlow(0)
-        override val openTradesSynced: StateFlow<Boolean> = MutableStateFlow(true)
-
-        override suspend fun getClosedTradesPaginated(
-            params: PaginationParams,
-            search: String?,
-            sortBy: TradeSort?,
-            outcomeFilter: TradeOutcomeFilter,
-            roleFilter: TradeRoleFilter,
-        ): Result<PaginatedResponse<ClosedTradeListItem>> = Result.success(PaginatedResponse(emptyList(), params.page, params.pageSize, 0L, 0))
-
-        override suspend fun takeOffer(
-            bisqEasyOffer: BisqEasyOfferVO,
-            takersBaseSideAmount: MonetaryVO,
-            takersQuoteSideAmount: MonetaryVO,
-            bitcoinPaymentMethod: String,
-            fiatPaymentMethod: String,
-            takeOfferStatus: MutableStateFlow<TakeOfferStatus?>,
-            takeOfferErrorMessage: MutableStateFlow<String?>,
-        ): Result<String> = takeOfferResult
-
-        override fun selectOpenTrade(tradeId: String) {}
-
-        override suspend fun rejectTrade(reason: AnalyticsEvent.Trade.InterruptReason): Result<Unit> = Result.success(Unit)
-
-        override suspend fun cancelTrade(reason: AnalyticsEvent.Trade.InterruptReason): Result<Unit> = Result.success(Unit)
-
-        override suspend fun closeTrade(): Result<Unit> = Result.success(Unit)
-
-        override suspend fun sellerSendsPaymentAccount(paymentAccountData: String): Result<Unit> = Result.success(Unit)
-
-        override suspend fun buyerSendBitcoinPaymentData(bitcoinPaymentData: String): Result<Unit> = Result.success(Unit)
-
-        override suspend fun sellerConfirmFiatReceipt(): Result<Unit> = Result.success(Unit)
-
-        override suspend fun buyerConfirmFiatSent(): Result<Unit> = Result.success(Unit)
-
-        override suspend fun sellerConfirmBtcSent(paymentProof: String?): Result<Unit> = Result.success(Unit)
-
-        override suspend fun btcConfirmed(): Result<Unit> = Result.success(Unit)
-
-        override suspend fun exportTradeDate(): Result<Unit> = Result.success(Unit)
-
-        override fun resetSelectedTradeToNull() {}
-    }
-
     @Test
     fun selectOfferToTake_fixedAmountSpec_noAmountRange() {
         // Arrange: USD market at $100,000/BTC
