+++ shared/presentation/src/commonMain/kotlin/network/bisq/mobile/presentation/trade/trade_detail/OpenTradePresenter.kt
@@ -127,6 +127,10 @@ class OpenTradePresenter(
         _selectedTrade.value = null
         _isInMediation.value = false
         _isTradeOutOfSync.value = false
+        // The chat row renders as soon as the trade is published, which happens before its collector
+        // runs, so these would carry the previous trade's numbers into the new one.
+        msgCount.value = 0
+        _lastChatMsg.value = null
 
         tradeJob?.cancel()
         tradeJob =
@@ -134,7 +138,7 @@ class OpenTradePresenter(
                 val currentTrade = tradesServiceFacade.selectOpenTradeWhenSynced(tradeId)
                 _selectedTrade.value = currentTrade
                 if (currentTrade == null) {
-                    log.w { "OpenTradePresenter.initialize found no trade $tradeId once the open trades synced - skipping flow collection" }
+                    log.w { "OpenTradePresenter.initialize found no trade ${tradeId.take(8)} once the open trades synced - skipping flow collection" }
                     _showTradeNotFoundDialog.value = true
                     return@launch
                 }
